Roam: Move the graph pass into the conversion
Block references were resolved after conversion, in the importer, by
matching lines of finished markdown. Nothing could test it, so the
recordings showed the refs unresolved.

Moved as it stands - the recordings that move are what it has been
doing all along, not a change to it. modifySourceBlockString loses an
async it never used.
