+++ src/formats/roam-json.ts
@@ -5,10 +5,8 @@ import { FormatImporter } from '../format-importer';
 import { i18n } from '../i18n';
 import { helpUrl } from '../constants';
 import { sanitizeFileName } from '../util';
-import { BlockInfo, RoamBlock, RoamPage } from './roam/models/roam-json';
-import { convertDateString, sanitizeFileNameKeepPath } from './roam/utils';
-import { RoamPageConverter } from './roam/convert';
-import { blockRefRegex, extractBlockReferenceUIDs } from './roam/block-refs';
+import { RoamPage } from './roam/models/roam-json';
+import { RoamGraphConverter } from './roam/graph';
 
 const regex = /{{pdf:|{{\[\[pdf|{{\[\[audio|{{audio:|{{video:|{{\[\[video/;
 const imageRegex = /https:\/\/firebasestorage(.*?)\?alt(.*?)\)/;
@@ -108,70 +106,7 @@ export class RoamJSONImporter extends FormatImporter {
 			const data = await file.readText();
 			const allPages = JSON.parse(data) as RoamPage[];
 
-			// PRE-PROCESS: map the blocks for easy lookup //
-			const [blockLocations, toPostProcess] = this.preprocess(allPages);
-
-			const markdownPages: Map<string, string> = new Map();
-			const pageUids: Map<string, string> = new Map();
-			for (const pageData of allPages) {
-				let pageName = convertDateString(sanitizeFileNameKeepPath(pageData.title), this.userDNPFormat).trim();
-				if (pageName === '') {
-					progress.reportFailed(pageData.uid, i18n.importer.roamJson.reasonEmptyTitle());
-					console.error('Cannot import data with an empty title', pageData);
-					continue;
-				}
-				const filename = `${graphFolder}/${pageName}.md`;
-
-				// if title option is enabled
-				const YAMLtitle = this.titleYAML ? pageData.title : '';
-
-				// if timestamp option is enabled
-				// set up numbers to pass, default to 0
-				let pageCreateTimestamp: number = 0;
-				let pageEditTimestamp: number = 0;
-				if (this.fileDateYAML) {
-					// get page creation time and update time
-					let pageCreateTime = pageData['create-time'];
-					let pageEditTime = pageData['edit-time'];
-
-					// type check both for numbers, set to 0 if there's a type mismatch
-					if (typeof pageCreateTime === 'number') {
-						pageCreateTimestamp = pageCreateTime;
-					}
-
-					if (typeof pageEditTime === 'number') {
-						pageEditTimestamp = pageEditTime;
-					}
-				}
-
-				// The attachment resolver needs the note's real parent for `.` and
-				// `./subfolder` settings, including Roam titles that create folders.
-				await this.createFolders(parseFilePath(filename).parent);
-				const converter = this.newConverter();
-				const markdownOutput = await converter.jsonToMarkdown(graphFolder, filename, pageData, '', false, YAMLtitle, pageCreateTimestamp, pageEditTimestamp);
-				markdownPages.set(filename, markdownOutput);
-				if (pageData.uid) pageUids.set(filename, pageData.uid);
-			}
-
-			// POST-PROCESS: fix block refs //
-			for (const callingBlock of toPostProcess.values()) {
-				const callingBlockStringScrubbed = await this.newConverter()
-					.roamMarkupScrubber(graphFolder, `${graphFolder}/${callingBlock.pageName}.md`, callingBlock.blockString, true);
-				const newCallingBlockReferences = await this.extractAndProcessBlockReferences(markdownPages, blockLocations, graphFolder, callingBlockStringScrubbed);
-
-				const callingBlockFilePath = `${graphFolder}/${callingBlock.pageName}.md`;
-				const callingBlockMarkdown = markdownPages.get(callingBlockFilePath);
-				if (callingBlockMarkdown) {
-					let lines = callingBlockMarkdown.split('\n');
-
-					let index = lines.findIndex((item: string) => item.contains('- ' + callingBlockStringScrubbed));
-					if (index !== -1) {
-						lines[index] = lines[index].replace(callingBlockStringScrubbed, newCallingBlockReferences);
-					}
-
-					markdownPages.set(callingBlockFilePath, lines.join('\n'));
-				}
-			}
+			const { pages: markdownPages, uids: pageUids } = await this.newGraphConverter(graphFolder, progress).convert(allPages);
 
 			// WRITE-PROCESS: create the actual pages //
 			const totalCount = markdownPages.size;
@@ -212,134 +147,20 @@ export class RoamJSONImporter extends FormatImporter {
 		return dailyPageFormat || 'YYYY-MM-DD';
 	}
 
-	private newConverter(): RoamPageConverter {
-		return new RoamPageConverter({
+	private newGraphConverter(graphFolder: string, progress: ImportContext): RoamGraphConverter {
+		return new RoamGraphConverter({
+			graphFolder,
 			userDNPFormat: this.userDNPFormat,
 			fileDateYAML: this.fileDateYAML,
 			titleYAML: this.titleYAML,
 			downloadAttachments: this.downloadAttachments,
 			downloadFirebaseFile: (blockText, folder) => this.downloadFirebaseFile(blockText, folder),
+			// The attachment resolver needs the note's real parent for `.` and
+			// `./subfolder` settings, including Roam titles that create folders.
+			prepareNote: async filename => void await this.createFolders(parseFilePath(filename).parent),
+			reportFailed: (id, reason) => progress.reportFailed(id, reason),
+			emptyTitleReason: i18n.importer.roamJson.reasonEmptyTitle(),
 		});
-	}
-
-	private preprocess(pages: RoamPage[]): Map<string, BlockInfo>[] {
-		// preprocess/map the graph so each block can be quickly found 
-		let blockLocations: Map<string, BlockInfo> = new Map();
-		let toPostProcessblockLocations: Map<string, BlockInfo> = new Map();
-		const userDNPFormat = this.userDNPFormat;
-
-		function processBlock(page: RoamPage, block: RoamBlock) {
-			if (block.uid) {
-				//check for roam DNP and convert to obsidian DNP
-				const dateObject = new Date(page.uid);
-				if (!isNaN(dateObject.getTime())) {
-					// The string can be converted to a Date object
-					const newPageTitle = convertDateString(page.title, userDNPFormat);
-					page.title = newPageTitle;
-				}
-
-				const info = {
-					pageName: sanitizeFileNameKeepPath(page.title),
-					blockString: block.string,
-				};
-
-				const containsBlockRefRegex = /.*?(\(\(.*?\)\)).*?/g;
-				if (containsBlockRefRegex.test(block.string)) {
-					toPostProcessblockLocations.set(block.uid, info);
-				}
-				blockLocations.set(block.uid, info);
-			}
-
-			if (block.children) {
-				for (let child of block.children) {
-					processBlock(page, child);
-				}
-			}
-		}
-
-		for (let page of pages) {
-			if (page.children) {
-				for (let block of page.children) {
-					processBlock(page, block);
-				}
-			}
-		}
-
-		return [blockLocations, toPostProcessblockLocations];
-	}
-
-
-	// setup to hold the newest and oldest timestamp value from a given page
-	newestTimestamp: number = 0;
-	oldestTimestamp: number = 0;
-
-
-	private async modifySourceBlockString(markdownPages: Map<string, string>, sourceBlock: BlockInfo, graphFolder: string, sourceBlockUID: string) {
-		if (!sourceBlock.blockString.endsWith('^' + sourceBlockUID)) {
-			const sourceBlockFilePath = `${graphFolder}/${sourceBlock.pageName}.md`;
-			let markdown = markdownPages.get(sourceBlockFilePath);
-
-			if (markdown) {
-				let lines = markdown.split('\n');
-
-				// Edit the specific line, for example, the 5th line.
-				let index = lines.findIndex((item: string) => item.contains('- ' + sourceBlock.blockString));
-				if (index !== -1) {
-					let newSourceBlockString = sourceBlock.blockString + ' ^' + sourceBlockUID;
-
-					// replace the line before updating sourceBlock
-					lines[index] = lines[index].replace(sourceBlock.blockString, newSourceBlockString);
-					sourceBlock.blockString = sourceBlock.blockString + ' ^' + sourceBlockUID;
-				}
-
-				markdownPages.set(sourceBlockFilePath, lines.join('\n'));
-			}
-		}
-	}
-
-	private async extractAndProcessBlockReferences(markdownPages: Map<string, string>, blockLocations: Map<string, BlockInfo>, graphFolder: string, inputString: string): Promise<string> {
-		const blockReferences = extractBlockReferenceUIDs(inputString);
-
-		// If there are no block references, return the input string as is
-		if (blockReferences.length === 0) {
-			return inputString;
-		}
-
-		// Asynchronously process each block reference
-		let processedBlocks: string[] = [];
-
-		for (const sourceBlockUID of blockReferences) {
-			try {
-				const sourceBlock = blockLocations.get(sourceBlockUID);
-
-				if (!sourceBlock) {
-					// no block with that uid exists
-					// most likely just double ((WITH_REGULAR_TEXT))
-					processedBlocks.push(sourceBlockUID);
-					continue;
-				}
-
-				// the source block string needs to be stripped of any page syntax or the alias won't work
-				let strippedSourceBlockString = sourceBlock.blockString.replace(/\[\[|\]\]/g, '');
-				// create the obsidian alias []()
-				let processedBlock = `[[${graphFolder}/${sourceBlock.pageName}#^${sourceBlockUID}|${strippedSourceBlockString}]]`;
-				// Modify the source block markdown page asynchronously so the new obsidian alias points to something
-				await this.modifySourceBlockString(markdownPages, sourceBlock, graphFolder, sourceBlockUID);
-
-				processedBlocks.push(processedBlock);
-			}
-			catch {
-				// no block with that uid exists
-				// most likely just double ((WITH_REGULAR_TEXT))
-				processedBlocks.push(sourceBlockUID);
-			}
-		}
-
-		// Replace the block references in the input string with the processed ones
-		let index = 0;
-		const processedString = inputString.replace(blockRefRegex, () => processedBlocks[index++]);
-
-		return processedString;
 	}
 
 	private async downloadFirebaseFile(line: string, sourcePath: string): Promise<string> {
