+++ tests/roam/expected/help-graph-excerpt/2021-10-07.md
@@ -1,1 +1,1 @@
-- ((0yr0CgkfA))
+- 0yr0CgkfA
