+++ tests/roam/expected/help-graph-excerpt/2023-02-10.md
@@ -8,21 +8,66 @@
                     - Congratulations! You have just created a new **Dropdown Template** 
                         - #.rm-g  #.rm-h 
                             - [ ] Edit the block below to change the template
-                                - [[help-graph-excerpt/RM/Component/Select]] an item from ((iYpGdGk0B)) [[roam/templates|RM/Component/Select]] an item from ((iYpGdGk0B)) [[roam/templates]]
+                                - [[help-graph-excerpt/RM/Component/Select]] an item from [[help-graph-excerpt/2023-02-10#^iYpGdGk0B|Dropdown List]] [[roam/templates|RM/Component/Select]] an item from [[help-graph-excerpt/2023-02-10#^iYpGdGk0B|Dropdown List ^iYpGdGk0B]] [[roam/templates]]
                                     - 
                             - [ ] Edit these blocks to change the items visible in the dropdown
-                                - Dropdown List
+                                - Dropdown List ^iYpGdGk0B
                                     - A
                                     - B
                                     - C
                     - [ ]  Apply the template #.rm-E 
                         - [ ] type `;;`   
-                        - {{((ovqdW0R_j)) 
+                        - {{[[help-graph-excerpt/2023-02-10#^ovqdW0R_j|RM/Component/Select an item from ((iYpGdGk0B)) roam/templates]] 
                         -  This will produce a custom Select Component like this one 
                             -  
-                            - built off this  ((iYpGdGk0B))
+                            - built off this  [[help-graph-excerpt/2023-02-10#^iYpGdGk0B|Dropdown List ^iYpGdGk0B]]
                     - This component is an experiment, and was built paritally as a demo of what you can do with  [[help-graph-excerpt/roam/render|roam/render]] - a feature of roam that allows you to build interactive components out of code blocks written *inside* your roam graph and referenced by the component in this format `` 
-                        - If you'd like to edit the code for the template - you can edit it ((pBiRi-Hba))
+                        - If you'd like to edit the code for the template - you can edit it [[help-graph-excerpt/2023-02-10#^pBiRi-Hba|```clojure
+(ns niki.dropdown
+  (:require
+    [clojure.string :as str]
+    [roam.block :as block]
+    [roam.datascript.reactive :as dr]))
+
+(defn replace-with-block-ref [block-uid opts-uid ref-uid]
+  (let [before (->> [:block/uid block-uid]
+                 (dr/pull [:block/string])
+                 (deref)
+                 :block/string)
+        pattern (re-pattern
+                  (str
+                    "\\{\\{"
+                    "(?:\\[\\[)?roam/render(?:\\]\\])?:" ;; roam/render
+                    "\\s*"
+                    "\\(\\(([a-zA-Z0-9-]+)\\)\\)"        ;; render code block
+                    "\\s*"
+                    "\\(\\((" opts-uid ")\\)\\)"         ;; options block
+                    "\\s*"
+                    "(\\(\\(([a-zA-Z0-9-]+)\\)\\))?"     ;; selected block
+                    "\\}\\}"))
+        after   (str/replace before pattern
+                  (str "{{roam/render: (($1)) (($2)) ((" ref-uid "))}}"))]
+    (block/update
+      {:block
+       {:uid block-uid 
+        :string after}})))
+
+(defn view-children-clickable [block-uid opts-uid selected-uid]
+  (let [children (->> [:block/uid opts-uid]
+                   (dr/pull [{:block/children [:block/uid :block/string]}])
+                   deref
+                   :block/children)]
+    [:select {:id opts-uid
+              :on-change #(replace-with-block-ref block-uid opts-uid (-> % .-target .-value))}
+     (for [c children]
+       [:option
+        {:value    (:block/uid c)
+         :selected (= (:block/uid c) selected-uid)
+         :on-change #(replace-with-block-ref (:block/string c))}
+        (:block/string c)])]))
+
+(defn main [{:keys [block-uid]} [_ opts-uid] & _ selected-uid]
+  [view-children-clickable block-uid opts-uid selected-uid])```]]
                             - ```clojure
                               (ns niki.dropdown
                                 (:require
