+++ tests/roam/expected/help-graph-excerpt/Secret Feature.md
@@ -1,2 +1,2 @@
-- ((V1RLPs0ai))
+- V1RLPs0ai
     - ![](https://twitter.com/example/status/1234567890?s=20)
